+++ apps/web/src/common/modules/widgets/_widget-fields/x-axis/__WidgetFieldXAxis.vue
@@ -7,7 +7,7 @@ import {
 import { PFieldGroup } from '@cloudforet/mirinae';
 import type { MenuItem } from '@cloudforet/mirinae/types/controls/context-menu/type';
 
-import WidgetFieldDropdownAndMax from '@/common/modules/widgets/_components/WidgetFieldDropdownAndMax.vue';
+import WidgetFieldDropdownAndMax from '@/common/modules/widgets/_components/__WidgetFieldDropdownAndMax.vue';
 import { useGranularityMenuItem } from '@/common/modules/widgets/_composables/use-granularity-menu-items';
 import {
     useWidgetOptionsComplexValidation,
